TA#24608 fix constraint on recording related to a product (#131)
Only block when the product is mapped from a product reference
or a catalog reference.

In such case, the product is expected to be linked to a recording.

If the revenue type is used to map the product, the constraint is not
raised. A revenue type is not bound to a specific recording.